pypuppetdb: Removing functionality to support for multiple API versions
Other project maintainers do not feel it beneficial to maintain multiple
API versions in future releases, citing more work than what it is work.
Moving all the functionality from api/v4.py to api/__init__.py.

Removing the api_version parameter from __init__.py.connect() since it is
no longer used.

Updating test_cases and the CHANGELOG to reflect changes
