+++ tests/test_connect.py
@@ -2,11 +2,6 @@ import pytest
 import pypuppetdb
 
 
-def test_connect_unknown_api_version():
-    with pytest.raises(pypuppetdb.errors.UnsupportedVersionError):
-        pypuppetdb.connect(api_version=1)
-
-
-def test_connect_api_v4():
-    puppetdb = pypuppetdb.connect(api_version=4)
+def test_connect_api():
+    puppetdb = pypuppetdb.connect()
     assert puppetdb.version == 'v4'
